Add EndpointPayload ABC and Specific Payload Dataclasses for API Requests (#208)
This PR adds 'payload' dataclasses for the API endpoints.

The EndpointPayload class is an Abstract Base Class, which defines the structure for all other payload classes. Payload classes are in charge of normalizing and validating the payload before it is used for a POST request. They also implement a .to_dict property for easy access to the correctly structured payload data as a python dictionary.