Improve response output (#220)
This PR introduces a way to remove `None` values from serialised
responses so they are more compact (and so that they mirror the REST API
response). Additionally, it changes how the data is returned to improve
clarity for users (by using clearer/more appropriate names):
- Remove the `.json` property. The rationale is that we weren't actually
returning a json string but a dictionary. Also, given that it was a
property, we couldn't take additional parameters to specify whether a
'compact' version should be returned.
- Add a `.to_dict()` method, which returns a dictionary. It takes a
parameter `exclude_none` (which defaults to `True`) which excludes any
keys which are None or empty lists*. I decided to call it `exclude_none`
rather than `compact` to mirror how Pydantic deals with the same issue
(model_dump optionally takes `exclude_none=True` in order to remove all
`None` values). This prepares us for a future move to Pydantic (instead
of Python dataclasses), as we've discussed.
- Add a `.to_json()` method, which returns a json string (via
`json.dumps`). It takes the same `exclude_none` parameter described
above.

Those two methods are implemented via a mixin class.