Add parents/ancestry methods (#231)
Fetch entity parents: A new Node method (`fetch_entity_parents`)
which uses `containedInPlace` to get the immediate parents for a given
entity or list of entities.

Fetch entity ancestry: A new Node method `fetch_entity_ancestry`
performs a parallel breadth-first traversal to construct the full
ancestry chain for an entity — all the way up to Earth — and returns it
in either flat or nested form.