Improvements suggested by kara (#234)
-removes `fetch_latest_observations` as it didn’t really serve a useful role
-removes`fetch_latest_observations_by_entity`
as it is basically the same as `fetch_observations_by_entity` but just
with a default of the latest date.
- renames`fetch_observations_by_entity` to `fetch_observations_by_entity_dcid`
- renames`get_observations_as_records` to `to_observation_records`. This name
more accurately reflects what the method does: it transforms nested
observation data and facet metadata into a list of flat, record-like
dictionaries suitable for analysis. This mirrors common data science
terminology.