Remove auto-flattening for unpack_arcs. (#244)
Currently the auto-flattening of unpack_arcs leads to the following
scenario:

Two properties are requested for a given entity but data only exists for
one of the properties. The REST api will not contain the property with
no data in the response, therefore we cannot tell **from only the
response** the difference between one property requested and one
property out of multiple properties requested contains data.

An example:
```
node_resp = dc_client.node.fetch(node_dcids="bio/APOE", expression="<-[encodesGene,variantID]")
node_resp.get_properties()
> {'bio/APOE': [Node(dcid='bio/AB035149.1', name='AB035149.1', provenanceId='dc/base/NCBI_Gene'...), ...] ...}
```

In the response from get_properties(), we can't tell if it's for
encodesGene or variantID.

A long term fix for this would be to store the original requested
properties in the NodeResponse object and only do this "autoflattening"
when there was one **requested** property. However, another
consideration is that many return types for a single method can cause a
lot confusion.