Add children / descendants methods  (#245)
Refactors the ascendency features in order to work for
children/descendants as well.

This creates convenient ways to fetch parents and children of given
entities. This can be done for the immediate cases, or recursively up
and down the graph.

`fetch_entity_children` takes one or more entity dcids and fetches its
children via the `containedInPlace`. It optionally accepts
`children_type` which gets only children of a specific type (e.g.,
'Country', 'State', 'IPCCPlace_50')

`fetch_entity_descendancy` is a similar method, but it does so via BFS
recursively. It is highly recommended to specify a type in that case, as
otherwise the search can become extremely large for entities further up
in the graph.