Fix bug in getting entity names (#252)
Fixes this error in fetch_entity_names:
https://screenshot.googleplex.com/Bm9BzSUnNYB87N9

There was a recent change that removes auto-flattening when unpacking
arcs (https://github.com/datacommonsorg/api-python/pull/244), but
fetch_entity_names was still assuming that values would be flattened.

Also updated the tests to use mock responses that are formatted the way
the response would actually come back as.