Removing surface_header_value validation (#277)
Completes b/453012979

We originally validated that the incoming surface header was either in
the form "mcp-<some version number" or "datagemma", to avoid public use
of this field that is supposed to only identify data commons surfaces
for mixer logs.

However, this makes it challenging to change the surface header or
introduce new surfaces, and doesn't allow release candidate versions of
the MCP server because the version number contains letters. I removed
this validation entirely and will instead validate that the surface is
an actual DC feature in my usage logs in mixer, where I can loosen
validation without releasing a new PyPi package.

(also updates version bump because of a previous change)